[opt](build) 2/x: Move int128 typedefs, RowsetId and FieldType out of olap_common.h
Three foundational entities lived in storage/olap_common.h, which drags
gen_cpp/Types_types.h, io_common, uid_util (boost/uuid) and friends into
every closure that needs any of them:

- int128_t/uint128_t -> core/extended_types.h. core/types.h reaches
  olap_common through binary_cast.hpp -> packed_int128.h purely for these
  two aliases, which put the storage layer underneath every TU in the
  codebase; util/coding.h had the same dependency for uint128_t.
- RowsetId -> storage/rowset_id.h (new), a dependency-free header
  (cstdint/functional/iosfwd/string) safe for core/column/column.h.
  Method bodies, the hex helpers dependency and MAX_ROWSET_ID/LOW_56_BITS
  move to storage/rowset_id.cpp; std::hash<RowsetId>::operator() is
  defined out of line so the header does not need hash_util.hpp.
- enum FieldType (+ the three field_is_*_type helpers) ->
  storage/field_type.h (new), so data-type headers can name storage cell
  types without the rest of olap_common.h.

olap_common.h includes the three light headers, so its own users are
unaffected. Includer fixes: packed_int128.h / util/coding.h /
data_type_number_serde.h now use extended_types.h; column.h /
column_nullable.h use rowset_id.h; data_type_decimal.h uses field_type.h;
data_type_ipv4.h and ipv4/ipv6 serde headers dropped a dead olap_common
include; decimal12.h carries its own power table instead of pulling
storage/utils.h; segment/rle_page.h takes field_type.h directly -- it
names FieldType as a non-type template parameter but reached the enum
only through util/coding.h, and its one live consumer (encoding_info.cpp)
masks that by including olap_common.h itself.

The closure-sweep natural-closure gate (300 first-party TUs, PCH
stripped) surfaced every TU that only compiled through the now-cut
transitive paths; fixed at the owning header: metadata_adder.h
(rowset_fwd), runtime_profile.h (cast_set), column_reader_cache.h
(io_common + olap_common), condition_cache.h (rowset_id),
python_client.h (olap_define for DISALLOW_COPY_AND_ASSIGN),
format_v2/file_reader.h and tracing_file_reader.h (io_common), four
data-type headers (field_type). function_encryption.cpp: the const-args
dispatch is now `if constexpr` so the arg_num==4 instantiation no longer
indexes col_const[4]/argument_columns[4] out of bounds (-Warray-bounds
fired once the include graph changed).

Verification: closure-sweep 300/300 clean vs p3-pre baseline;
compile_bench full rebuild + doris_be link green at -j6; BE UT Debug
build green, 31 tests / 6 suites passed (RowsetId behavior:
IdManagerTest, IdFileMapTest, StorageResourceTest, TestRowIdConversion,
TxnManagerTest; registration contract: BinaryArithmeticRegistrationTest).
