+++ be/src/core/column/column.h
@@ -36,7 +36,7 @@
 #include "core/string_ref.h"
 #include "core/typeid_cast.h"
 #include "core/types.h"
-#include "storage/olap_common.h"
+#include "storage/rowset_id.h"
 
 namespace doris {
 class SipHash;
