+++ be/src/core/data_type/data_type_fixed_length_object.h
@@ -31,6 +31,7 @@
 #include "core/data_type_serde/data_type_string_serde.h"
 #include "core/field.h"
 #include "core/types.h"
+#include "storage/field_type.h"
 
 namespace doris {
