+++ be/src/core/data_type_serde/data_type_number_serde.h
@@ -29,10 +29,10 @@
 #include "core/data_type/data_type.h"
 #include "core/data_type/define_primitive_type.h"
 #include "core/data_type_serde/data_type_serde.h"
+#include "core/extended_types.h"
 #include "core/field.h"
 #include "core/string_ref.h"
 #include "core/types.h"
-#include "storage/olap_common.h"
 
 namespace doris {
 class JsonbOutStream;
