+++ be/src/core/data_type_serde/data_type_string_serde.h
@@ -28,6 +28,7 @@
 #include "core/column/column_string.h"
 #include "core/data_type_serde/data_type_serde.h"
 #include "core/types.h"
+#include "storage/field_type.h"
 
 namespace doris {
 class PValues;
